+++ Runtime/Scripts/Utilities/ObjectUtility.cs
@@ -10,23 +10,35 @@ using UnityEditor;
 namespace Unity.FilmInternalUtilities {
 internal static class ObjectUtility {
 
+#if UNITY_2020_3_OR_NEWER
+
+    internal static IEnumerable<T> FindSceneComponents<T>(bool includeInactive = true) where T: UnityEngine.Component {
+        foreach (T comp in Object.FindObjectsOfType<T>(includeInactive)) {
+            yield return comp;
+        }
+    }
+#else
     internal static IEnumerable<T> FindSceneComponents<T>() where T: UnityEngine.Component {
         foreach (Object o in Resources.FindObjectsOfTypeAll(typeof(T))) {
             T comp = (T) o;
             Assert.IsNotNull(comp);
-            GameObject go   = comp.gameObject;
+            GameObject go = comp.gameObject;
 
             if (!(go.hideFlags == HideFlags.NotEditable || go.hideFlags == HideFlags.HideAndDontSave)
 #if UNITY_EDITOR                    
                 && !EditorUtility.IsPersistent(comp.transform.root.gameObject)
 #endif                    
-            ) 
+               ) 
             {
                 yield return comp;
                     
             }
         }
     }
+    
+#endif
+
+
 
 //----------------------------------------------------------------------------------------------------------------------